# apps/nest-server-orm/src/assets/swagger.json
{
...
    "openapi": "3.0.0",
    "servers": [
        {
        "url": "./",
        "description": "Events API"
        }
    ],
    "schemes": [
      "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
...
    "paths": {
...
        "/users": {
            "get": {
                "summary": "List all users",
                "tags": ["Users"],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserDetailView"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new user",
                "tags": ["Users"],
                "requestBody": {
                    "description": "New user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDetailView"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "Retrieve a user",
                "tags": ["Users"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDetailView"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a user",
                "requestBody": {
                    "description": "User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDetailView"
                            }
                        }
                    }
                },
                "tags": ["Users"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "delete": {
                "summary": "Delete an user",
                "tags": ["Users"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": "true"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserDetailView": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "username": {
                        "type": "string",
                        "example": "An Interesting Event"
                    }
                }
            }
        },
...
    }
}